refactor : 자동차 이름을 나타내는 CarName 정의
- Car에 있던 이름 관련 테스트를 CarName으로 이동
- Car에서는 CarName을 통해서만 이름에 접근
